requirements.txt
fastapi==0.127.1
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
psycopg2-binary==2.9.11
pgvector==0.4.2
numpy==2.3.5
pandas==2.3.3
faiss-cpu==1.13.1
sentence-transformers==5.1.2
torch==2.9.1
tqdm==4.67.1
scikit-learn==1.7.2
scipy==1.16.3
